pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.50.0"
description = ""
authors = ["tomilinBV <qwertere97@mail.ru>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8.5"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
